Add files via upload
This is a power bi visual dashboard of "global superstore" showing the sales through various sectors of the sales. This visualization of sales data provides insights of sales of the store at various geographic divisions and it also gives a deeper understanding on sales based on the shipping cost trends.